py-gpu-inference/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coral-gpu-inference"
version = "0.1.0"
description = "Production-grade gRPC GPU embedding inference server with token-bucket batching, backpressure, and torch.compile support."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Coral Bricks AI", email = "eng@coralbricks.ai"},
]
keywords = ["gpu", "inference", "embeddings", "grpc", "pytorch", "transformers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "torch>=2.5.0",
    "transformers>=4.40.0",
    "safetensors>=0.4.0",
    "numpy>=1.24.0",
    "huggingface-hub>=0.20.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.28.0"]
cloudwatch = ["boto3>=1.28.0"]
flash-attn = ["flash-attn>=2.5.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21.0"]

[project.urls]
Homepage = "https://github.com/Coral-Bricks-AI/coral-ai"
Repository = "https://github.com/Coral-Bricks-AI/coral-ai"
Issues = "https://github.com/Coral-Bricks-AI/coral-ai/issues"

[tool.setuptools.packages.find]
include = ["coral_gpu_inference*"]
